server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.0
pydantic[email]>=2.4.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
motor>=3.3.0
pymongo>=4.6.0
